requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
SpeechRecognition==3.10.1
gTTS==2.5.0
langdetect==1.0.9
pydub==0.25.1
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0